# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/containers": {
      "post": {
        "operationId": "ContainersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContainerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Container created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Place or Room not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new container",
        "tags": [
          "Containers"
        ]
      },
      "get": {
        "operationId": "ContainersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Containers retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all containers",
        "tags": [
          "Containers"
        ]
      }
    },
    "/containers/{id}": {
      "get": {
        "operationId": "ContainersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Container ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Container retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Container not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get container by ID",
        "tags": [
          "Containers"
        ]
      },
      "patch": {
        "operationId": "ContainersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Container ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContainerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Container updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Container, Place or Room not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update container",
        "tags": [
          "Containers"
        ]
      },
      "delete": {
        "operationId": "ContainersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Container ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Container deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Container not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete container",
        "tags": [
          "Containers"
        ]
      }
    },
...
  },
  "info": {
    "title": "ShelfSpot API",
    "description": "API for ShelfSpot application with JWT Authentication",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateContainerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Storage Box",
            "description": "Container name"
          },
          "icon": {
            "type": "string",
            "example": "box",
            "description": "Container icon"
          },
          "placeId": {
            "type": "number",
            "example": 1,
            "description": "Place ID"
          },
          "roomId": {
            "type": "number",
            "example": 1,
            "description": "Room ID"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateContainerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Storage Box",
            "description": "Container name"
          },
          "icon": {
            "type": "string",
            "example": "box",
            "description": "Container icon"
          },
          "placeId": {
            "type": "number",
            "example": 1,
            "description": "Place ID"
          },
          "roomId": {
            "type": "number",
            "example": 1,
            "description": "Room ID"
          }
        }
      },
...
    }
  }
}